moves message fixtures to a common helper file
