+++ insecure/attacknetwork/corruptedConnector_test.go
@@ -1,4 +1,4 @@
-package attacknetwork_test
+package attacknetwork
 
 import (
 	"context"
@@ -7,57 +7,30 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/onflow/flow-go/insecure/attacknetwork"
-	"github.com/onflow/flow-go/insecure/corruptible"
-	mockinsecure "github.com/onflow/flow-go/insecure/mock"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/irrecoverable"
-	"github.com/onflow/flow-go/network/codec/cbor"
 	"github.com/onflow/flow-go/utils/unittest"
 )
 
 func TestConnectorHappyPath(t *testing.T) {
-	withCorruptibleConduitFactory(t, func(identity flow.Identity, factory *corruptible.ConduitFactory) {
-		corruptedIds := flow.IdentityList{&identity}
-		withAttackNetwork(t, corruptedIds, func(attackNetwork *attacknetwork.AttackNetwork) {
+	withMockCorruptibleConduitFactory(t, func(corruptedId flow.Identity, factory *mockCorruptibleConduitFactory) {
+		connector := NewCorruptedConnector(flow.IdentityList{&corruptedId})
 
-		})
+		go func() {
+			msg := <-factory.attackerRegMsg
+
+		}()
+
+		connection, err := connector.Connect(context.Background(), corruptedId.NodeID)
+		require.NoError(t, err)
+
+		connection.SendMessage()
+
 	})
 }
 
-func withAttackNetwork(t *testing.T, corruptedIds flow.IdentityList, run func(*attacknetwork.AttackNetwork)) {
-	codec := cbor.NewCodec()
-	o := &mockinsecure.AttackOrchestrator{}
-	connector := attacknetwork.NewCorruptedConnector(corruptedIds)
-	attackNetwork, err := attacknetwork.NewAttackNetwork(unittest.Logger(), "localhost:0", codec, o, connector, corruptedIds)
-	require.NoError(t, err)
-
-	// life-cycle management of attackNetwork.
-	ctx, cancel := context.WithCancel(context.Background())
-	attackNetworkCtx, errChan := irrecoverable.WithSignaler(ctx)
-	go func() {
-		select {
-		case err := <-errChan:
-			t.Error("attackNetwork startup encountered fatal error", err)
-		case <-ctx.Done():
-			return
-		}
-	}()
-
-	// starts corruptible conduit factory
-	attackNetwork.Start(attackNetworkCtx)
-	unittest.RequireCloseBefore(t, attackNetwork.Ready(), 1*time.Second, "could not start attack network on time")
-
-	run(attackNetwork)
-
-	// terminates attackNetwork
-	cancel()
-	unittest.RequireCloseBefore(t, attackNetwork.Done(), 1*time.Second, "could not stop attack network on time")
-}
-
-func withCorruptibleConduitFactory(t *testing.T, run func(flow.Identity, *corruptible.ConduitFactory)) {
-	codec := cbor.NewCodec()
-	corruptedIdentity := unittest.IdentityFixture()
+func withMockCorruptibleConduitFactory(t *testing.T, run func(flow.Identity, *mockCorruptibleConduitFactory)) {
+	corruptedIdentity := unittest.IdentityFixture(unittest.WithAddress("localhost:0"))
 
 	// life-cycle management of attackNetwork.
 	ctx, cancel := context.WithCancel(context.Background())
@@ -65,12 +38,13 @@ func withCorruptibleConduitFactory(t *testing.T, run func(flow.Identity, *corrup
 	go func() {
 		select {
 		case err := <-errChan:
-			t.Error("attackNetwork startup encountered fatal error", err)
+			t.Error("attack network startup encountered fatal error", err)
 		case <-ctx.Done():
 			return
 		}
 	}()
-	ccf := corruptible.NewCorruptibleConduitFactory(unittest.Logger(), flow.BftTestnet, corruptedIdentity.NodeID, codec, "localhost:0")
+
+	ccf := newMockCorruptibleConduitFactory("localhost:5000")
 
 	// starts corruptible conduit factory
 	ccf.Start(ccfCtx)
